Moved improvepage into /Lesson, some css
